flashing messages while logging out

+++ src/App.js
@@ -1,10 +1,15 @@
-import { BrowserRouter as Router, Switch, Route, useHistory } from "react-router-dom";
-import React from 'react'
+import { 
+  BrowserRouter as Router, 
+  Switch, 
+  Route, 
+  useHistory,
+  Redirect
+} from "react-router-dom";
+import EventEmitter from "events";
 import AppDashboard from './screens/App';
 import Home from "./screens/Home";
 import Signin from "./screens/Signin";
 import Signup from "./screens/Signup";
-import  EventEmitter from "events";
 
 export const event = new EventEmitter()
 
@@ -22,11 +27,17 @@ const App = () => {
         <Route exact path="/login" component={Signin} />
         <Route exact path="/register" component={Signup} />
         <Route exact path="/logout" component={() => {
-          localStorage.clear()
-          setTimeout(() => {
-            window.flash('Logged out successfully', 'success')
-          }, 100)
-          history.push('/')
+          if (!localStorage.getItem('userId')) {
+            setTimeout(() => {
+              window.flash('You were not logged in', 'warning')
+            }, 100)
+          } else {
+            localStorage.clear()
+            setTimeout(() => {
+              window.flash('Logged out successfully', 'success')
+            }, 100)
+          }
+          return <Redirect to='/' />
         }} />
       </Switch>
     </Router>
